requirements.txt
streamlit
selenium
pandas
mysql-connector-python
pymysql
streamlit-option-menu
webdriver-manager